full/events_analysis/requirements.txt
pyspark
python-dateutil
numpy
pandas
scikit-learn
torch
json
time
networkx
kafka-python
requests
argparse
logging
datetime